refactor: extract AppLifecycleNotificationBinder.swift from AppPresentationState.swift
Splits the `AppLifecycleNotificationBinder` helper (subscribes to
NSApplication activation notifications and forwards them to
callbacks) out of AppPresentationState.swift into its own file.

Byte-identical move of lines 22-46. Both files import AppKit + Combine +
Foundation.

AppPresentationState.swift: 76 → 50 lines. Tests: 667.
