+++ Sources/BugNarrator/Utilities/AppPresentationState.swift
@@ -20,32 +20,6 @@ final class ObservableObjectChangeForwarder {
 }
 
 @MainActor
-final class AppLifecycleNotificationBinder {
-    private let notificationCenter: NotificationCenter
-    private var cancellables = Set<AnyCancellable>()
-
-    init(notificationCenter: NotificationCenter = .default) {
-        self.notificationCenter = notificationCenter
-    }
-
-    func bind(didBecomeActive: @escaping () -> Void, willTerminate: @escaping () -> Void) {
-        cancellables.removeAll()
-
-        notificationCenter.publisher(for: NSApplication.didBecomeActiveNotification)
-            .sink { _ in
-                didBecomeActive()
-            }
-            .store(in: &cancellables)
-
-        notificationCenter.publisher(for: NSApplication.willTerminateNotification)
-            .sink { _ in
-                willTerminate()
-            }
-            .store(in: &cancellables)
-    }
-}
-
-@MainActor
 final class AppPresentationState: ObservableObject {
     @Published private(set) var status: AppStatus
     @Published private(set) var currentError: AppError?
